macosx: improved mouse emulation
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@25414 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/macosx/cocoa/menu/vicemachinecontroller.h
@@ -126,7 +126,7 @@
 
 // ----- Mouse & Lightpen -----
 // mouse moved
--(void)mouseMoveToX:(int)x andY:(int)y;
+-(void)mouseMoveToX:(float)x andY:(float)y;
 // mouse pressed
 -(void)mouseButton:(BOOL)left withState:(BOOL)pressed;
 // lightpen update
